+++ CmdletHelpGenerator/Program.cs
@@ -64,7 +64,7 @@ namespace OfficeDevPnP.PowerShell.CmdletHelpGenerator
                     var category = string.Empty;
                     var attrs = t.GetCustomAttributes();
                     var examples = new List<CmdletExampleAttribute>();
-
+                    var relatedLinks = new List<CmdletRelatedLinkAttribute>();
                     //System.Attribute.GetCustomAttributes(t); 
 
                     // Get info from attributes
@@ -90,6 +90,11 @@ namespace OfficeDevPnP.PowerShell.CmdletHelpGenerator
                         {
                             var a = (CmdletExampleAttribute)attr;
                             examples.Add(a);
+                        }
+                        if (attr is CmdletRelatedLinkAttribute)
+                        {
+                            var l = (CmdletRelatedLinkAttribute)attr;
+                            relatedLinks.Add(l);
                         }
                     }
 
@@ -291,6 +296,24 @@ namespace OfficeDevPnP.PowerShell.CmdletHelpGenerator
                         exampleCount++;
                     }
                     commandElement.Add(examplesElement);
+
+                    // Related links
+                    var relatedLinksElement = new XElement(maml + "relatedLinks");
+                    relatedLinks.Insert(0, new CmdletRelatedLinkAttribute() { Text = "Online Version", Url = "http://aka.ms/officedevpnp" });
+
+                    foreach (var link in relatedLinks)
+                    {
+                        var navigationLinksElement = new XElement(maml + "navigationLink");
+                        var linkText = new XElement(maml + "linkText");
+                        linkText.Value = link.Text + ":";
+                        navigationLinksElement.Add(linkText);
+                        var uriElement = new XElement(maml + "uri");
+                        uriElement.Value = link.Url;
+                        navigationLinksElement.Add(uriElement);
+
+                        relatedLinksElement.Add(navigationLinksElement);
+                    }
+                    commandElement.Add(relatedLinksElement);
 
                     // Markdown from CmdletInfo
                     if (generateMarkdown)