+++ Commands/Fields/AddTaxonomyField.cs
@@ -4,52 +4,86 @@ using Microsoft.SharePoint.Client;
 using OfficeDevPnP.Core.Entities;
 using OfficeDevPnP.PowerShell.CmdletHelpAttributes;
 using OfficeDevPnP.PowerShell.Commands.Base.PipeBinds;
+using Microsoft.SharePoint.Client.Taxonomy;
 
 namespace OfficeDevPnP.PowerShell.Commands
 {
     [Cmdlet(VerbsCommon.Add, "SPOTaxonomyField")]
     [CmdletHelp("Adds a taxonomy field to a list or as a site column.",
         Category = CmdletHelpCategory.Fields)]
+    [CmdletExample(
+        Code = @"PS:> Add-SPOTaxonomyField -DisplayName ""Test"" -InternalName ""Test"" -TermSetPath ""TestTermGroup|TestTermSet""",
+        Remarks = @"Adds a new taxonomy field called ""Test"" that points to the TestTermSet which is located in the TestTermGroup",
+        SortOrder = 1)]
     public class AddTaxonomyField : SPOWebCmdlet
     {
-        [Parameter(Mandatory = false)]
+        [Parameter(Mandatory = false, ParameterSetName = ParameterAttribute.AllParameterSets)]
         public ListPipeBind List;
 
-        [Parameter(Mandatory = true)]
+        [Parameter(Mandatory = true, ParameterSetName = ParameterAttribute.AllParameterSets)]
         public string DisplayName;
 
-        [Parameter(Mandatory = true)]
+        [Parameter(Mandatory = true, ParameterSetName = ParameterAttribute.AllParameterSets)]
         public string InternalName;
 
-        [Parameter(Mandatory = true)]
+        [Parameter(Mandatory = true, ParameterSetName = "Path")]
         public string TermSetPath;
 
-        [Parameter(Mandatory = false)]
+        [Parameter(Mandatory = false, ParameterSetName = "Id")]
+        public GuidPipeBind TaxonomyItemId;
+
+        [Parameter(Mandatory = false, ParameterSetName = "Path")]
         public string TermPathDelimiter = "|";
 
-        [Parameter(Mandatory = false)]
+        [Parameter(Mandatory = false, ParameterSetName = ParameterAttribute.AllParameterSets)]
         public string Group;
 
-        [Parameter(Mandatory = false)]
+        [Parameter(Mandatory = false, ParameterSetName = ParameterAttribute.AllParameterSets)]
         public GuidPipeBind Id = new GuidPipeBind();
 
-        [Parameter(Mandatory = false)]
+        [Parameter(Mandatory = false, ParameterSetName = ParameterAttribute.AllParameterSets)]
         public SwitchParameter AddToDefaultView;
 
-        [Parameter(Mandatory = false)]
+        [Parameter(Mandatory = false, ParameterSetName = ParameterAttribute.AllParameterSets)]
         public SwitchParameter MultiValue;
 
-        [Parameter(Mandatory = false)]
+        [Parameter(Mandatory = false, ParameterSetName = ParameterAttribute.AllParameterSets)]
         public SwitchParameter Required;
 
-        [Parameter(Mandatory = false)]
+        [Parameter(Mandatory = false, ParameterSetName = ParameterAttribute.AllParameterSets)]
         public AddFieldOptions FieldOptions = AddFieldOptions.DefaultValue;
 
 
         protected override void ExecuteCmdlet()
         {
+            TaxonomyItem taxItem = null;
             Field field;
-            var termSet = ClientContext.Site.GetTaxonomyItemByPath(TermSetPath, TermPathDelimiter);
+            if (ParameterSetName == "Path")
+            {
+                taxItem = ClientContext.Site.GetTaxonomyItemByPath(TermSetPath, TermPathDelimiter);
+            }
+            else
+            {
+                var taxSession = ClientContext.Site.GetTaxonomySession();
+                var termStore = taxSession.GetDefaultKeywordsTermStore();
+                try
+                {
+                    taxItem = termStore.GetTermSet(TaxonomyItemId.Id);
+                }
+                catch
+                {
+                    try
+                    {
+                        taxItem = termStore.GetTerm(TaxonomyItemId.Id);
+                    }
+                    catch
+                    {
+                        throw new Exception(string.Format("Taxonomy Item with Id {0} not found", TaxonomyItemId.Id));
+                    }
+                }
+                taxItem.EnsureProperty(t => t.Id);
+            }
+
             Guid id = Id.Id;
             if (id == Guid.Empty)
             {
@@ -62,7 +96,7 @@ namespace OfficeDevPnP.PowerShell.Commands
                 InternalName = InternalName,
                 DisplayName = DisplayName,
                 Group = Group,
-                TaxonomyItem = termSet,
+                TaxonomyItem = taxItem,
                 MultiValue = MultiValue,
                 Required = Required,
                 AddToDefaultView = AddToDefaultView
@@ -71,9 +105,6 @@ namespace OfficeDevPnP.PowerShell.Commands
             if (List != null)
             {
                 var list = List.GetList(SelectedWeb);
-
-              
-
                 field = list.CreateTaxonomyField(fieldCI);
             }
             else